Update PluginInstaller.cs
  - Fixed Error in system that caused crash when a user removes a plugin manually, then goes into Godot Manager, and unchecks the plugin.

Fixes #23

+++ libs/util/PluginInstaller.cs
@@ -167,6 +167,9 @@ public class PluginInstaller : Object
 	}
 
 	void DeleteImports(string path) {
+		if (!Directory.Exists(path.GetOSDir().NormalizePath()))
+			return;
+		
 		foreach(string file in Directory.EnumerateFiles(path)) {
 			if (file.EndsWith(".import"))
 				SFile.Delete(file);
